requirements.txt
addict==2.4.0
annoy==1.17.3
asttokens==2.2.1
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.1.6
cmake==3.27.0
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.0
debugpy==1.6.7
decorator==5.1.1
docker-pycreds==0.4.0
executing==1.2.0
filelock==3.12.2
fonttools==4.41.1
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
huggingface-hub==0.16.4
idna==3.4
imagecorruptions==1.1.2
imageio==2.31.1
importlib-metadata==6.8.0
ipykernel==6.25.0
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
lazy_loader==0.3
lit==16.0.6
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-dataclass==8.5.14
matplotlib==3.5.1
matplotlib-inline==0.1.6
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.5.7
networkx==3.1
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnx==1.12.0
opencv-python==4.6.0.66
packaging==23.1
pandas==1.4.2
parso==0.8.3
pathlib==1.0.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.9.1
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.20.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
regex==2023.6.3
requests==2.27.1
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.11.2
sentry-sdk==1.28.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
sklearn==0.0.post7
smmap==5.0.0
stack-data==0.6.2
sympy==1.12
terminaltables==3.1.10
thop==0.1.1.post2207130030
threadpoolctl==3.2.0
tifffile==2023.7.18
tokenizers==0.13.3
tomli==2.0.1
torch==1.12.1
torchvision==0.13.1
tornado==6.3.2
tqdm==4.64.0
traitlets==5.9.0
transformers==4.31.0
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==1.26.16
wandb==0.13.6
wcwidth==0.2.6
yapf==0.40.1
zipp==3.16.2

